% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tech_reps.R
\name{combine_tech_reps}
\alias{combine_tech_reps}
\title{Combine technical replicates using top 3 methods}
\usage{
combine_tech_reps(reps, normalize = TRUE, match_peps = TRUE,
  relabel = FALSE)
}
\arguments{
\item{rep1}{}

\item{rep2}{}
}
\description{
Combine technical replicates using top 3 methods
}

