% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_peptides.R
\name{map_peptides}
\alias{map_peptides}
\title{Map peptides to their locations within a protein}
\usage{
map_peptides(msf_file, min_conf = "High", prot_regex = "")
}
\arguments{
\item{min_conf}{}
}
\value{
A dataframe containing start and stop positions (relative to the parent protein sequence) for each peptide in the database.
}
\description{
Takes a thermo MSF file and uses `stringr` to find the location of peptides within their corresponding protein sequences.
}
\examples{
map_peptides("mythermofile.msf")
}

