% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tech_reps.R
\name{merge_top_peptides}
\alias{merge_top_peptides}
\title{Average top three peptides from multiple replicates}
\usage{
merge_top_peptides(df, num_reps)
}
\arguments{
\item{df}{}
}
\description{
Average top three peptides from multiple replicates
}

