\name{parsemsf-package}
\alias{parsemsf-package}
\alias{parsemsf}
\docType{package}
\title{
\packageTitle{parsemsf}
}
\description{
\packageDescription{parsemsf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{parsemsf}
\packageIndices{parsemsf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{parsemsf}

Maintainer: \packageMaintainer{parsemsf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
