% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{parsemsf_example}
\alias{parsemsf_example}
\title{Get path to parsemsf example}
\usage{
parsemsf_example(path)
}
\arguments{
\item{path}{Name of file}
}
\description{
From Hadley Wickham's readr package. parsemsf comes bundled with a number of sample
files in its \code{inst/extdata} directory. This function make them easy to access
}
\examples{
parsemsf_example("test_db.msf")
}
\keyword{internal}
