\name{protIDsplitter}
\alias{protIDsplitter}
\title{Internal Function for splitting protein name}
\description{
}
\usage{
protIDsplitter(cur.str)
}
\arguments{
  \item{cur.str}{
}
}
\value{

}

\author{
Ashoka Polpitiya
}
\seealso{
  \code{\link{protModsTable}}, \code{\link{scoreTable}}, \code{\link{mtzTable}}, \code{\link{protAnnotations}}, \code{\link{protModsTable}}, \code{\link{aucTable}}
}
\keyword{misc}

