\name{scoreTable}
\alias{scoreTable}
\title{Peptide ID scores from the search engine}
\usage{
scoreTable(msfFile, minConf = "High")
}
\arguments{
  \item{msfFile}{
}
  \item{minConf}{
}
}
\value{
}
\author{
Ashoka Polpitiya
}
\seealso{
  \code{\link{protModsTable}}, \code{\link{mtzTable}}, \code{\link{protAnnotations}}, \code{\link{protModsTable}}, \code{\link{aucTable}}
}
\examples{
\dontrun{M <- scoreTable(msfFile="D:/Research/R/msfRead/test_dimethyl.msf", minConf="High")}
}
\keyword{misc}
