% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_auc_table.R
\name{make_auc_table}
\alias{make_auc_table}
\title{Title}
\usage{
make_auc_table(msf_file, min_conf = "High", as_df = TRUE)
}
\arguments{
\item{msf_file}{}

\item{min_conf}{}

\item{as_df}{}
}
\description{
Title
}

