% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pep_table.R
\name{make_pep_table}
\alias{make_pep_table}
\title{Make a table of peptides}
\usage{
make_pep_table(msf_file, min_conf = "High")
}
\arguments{
\item{msf_file}{}

\item{min_conf}{}
}
\description{
Make a table of peptides
}

