% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pep_table.R
\name{make_pep_table}
\alias{make_pep_table}
\title{Make a data frame of peptides from a thermo MSF file}
\usage{
make_pep_table(msf_file, min_conf = "High",
  prot_regex = "^>([a-zA-Z0-9._]+)\\\\b", collapse = TRUE)
}
\arguments{
\item{msf_file}{A file path to a thermo MSF file.}

\item{min_conf}{"High", "Medium", or "Low". The mininum peptide confidence level to retrieve from MSF file.}

\item{prot_regex}{Regular expression where the first group matches a protein name or ID from the protein description. Regex must contain ONE group. The protein description is typically generated from a fasta reference file that was used for the database search.}

\item{collapse}{If TRUE, peptides that match to multiple protein sequences are collapsed into a single row with multiple protein IDs in the `Proteins` column, separated by a ";".}
}
\value{
A data frame of all peptides above the confidence cut-off from a thermo MSF file.

\item{PeptideID}{a unique peptide ID}
\item{SpectrumID}{?}
\item{Sequence}{amino acid sequence (does not show post-translational modifications)}
\item{Proteins}{protein name parsed from protein description, typically generated from database search reference file}
\item{ProteinID}{unique protein ID to which this peptide maps}
\item{PEP}{?}
\item{q-Value}{?}
}
\description{
Make a data frame of peptides from a thermo MSF file
}
\examples{
make_pep_table("mythermofile.msf")

# Retrieve full protein description

make_pep_table("mythermofile.msf", prot_regex = "")

# ...which is also equivalent to...

make_pep_table("mythermofile.msf", prot_regex = "^(.+)$")

}

