% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_table.R
\name{make_score_table}
\alias{make_score_table}
\title{Make a table of scores}
\usage{
make_score_table(msf_file, min_conf = "High")
}
\arguments{
\item{msf_file}{}

\item{min_conf}{}
}
\description{
Make a table of scores
}

